Move helper to sub-directory
