+++ src/services/helpers/scopePermissions/hooks/rejectQueryingForOtherUsers.js
@@ -1,5 +1,5 @@
 const { Forbidden, BadRequest } = require('@feathersjs/errors');
-const { resolveScope } = require('./resolveScope');
+const { resolveScope } = require('./util/resolveScope');
 
 const rejectQueryingOtherUsers = (context) => {
 	if (context.params === undefined || context.params.account === undefined) {
