+++ src/services/helpers/scopePermissions/hooks/util/resolveScope.js
